Ignore section tasks
Section tasks never represent a story. They should not be included in
any story based calculation. This is determined by looking at the
`resourece_subtype` of the task. Only `default_task` types should be
included. This pushes some `Maybe` behavior into all the executables,
but it is an easy situation to overcome.