Ensure trailing newline in output
The fold was placing a leading newline, desired to separate the totals
from the logging above. It did not include a trailing newline, which
messes up my prompt when the tool is run.

We could explicitly do that, but it was easier to switch the fold to
place trailing newlines (including a final one) and then have an
explicitly empty first element in our list of lines.