Add cycle time measure
Product level cycle time is an important metric to keep an eye on. How
quickly we are able to iterate on ideas and ship them to production for
validation says a lot about how nimble we are as an organization. Large
regressions in these numbers may indicate shifts in business priority,
breakdowns in process or other maladaptations.

This executable is an experimental tool to measure that from Asana task
lifetime. This does not capture the entire cycle time of a product idea
from inception to shipping, but it gives us a rough estimate from
business processes we can observe and measure.