Replace putEnumField with putCustomFields
By enhancing the CustomField type enough to carry information necessary
for updating values of that type, we can make the PUT function clearer.

In the current usage, we still construct things from static values, but
in a future usage we will use the CustomFields read from a GET to build
the PUT, this will accomplish the required fieldId update without having
to hard-code more fieldIds.