Fix link rendering (#15)
Links don't render properly in the title field in planning poker for
some reason, so I dropped the link down into the summary. This actually
simplifies some things because we were having to do some string
manipulation to extract the text from a link to recover the original
issue key from the exported game.