Various close-iteration improvements (#18)
* Various close-iteration improvements

This commit works with a new version of the Red/Green Iteration template
that has carry in (from previous iteration) and carry out (into next
interation) set as different fields. The calculation for velocity as it
stood was wrong, as it frequently conflated these two concepts.

This commit fixes the calculation of points completed using these two
fields in a backwards-compatible way (for now) and also adds it to a new
field in the project template "points completed" which marks how much of
the cost of each task was completed within the given project. The sum of
'points completed' is the velocity for the iteration.

I plan to add some tooling to start-iteration to swap carry-out to
carry-in and add an 'effective cost' field for each task which marks how
many points are being committed to for the current project for each
task. These stats will give us a much clearer picture of the velocity
for each iteration.

* Rollback start-iteration updates (will deal with those later)

* Use `implies` in storyStats1

* Lint

* Style updates