Update start-iteration workflow
This updates the `start-iteration` workflow with the new structure of
the R/G asana templates, which includes:

1. `carry in`
2. `carry out`
3. `commitment`

This also adds a couple of new warnings as well as the ability to swap
'carry out' for 'carry in', which is necessary at the beginning of the
database. It also allows adding `commitment` values so we can easily
track the target velocity of each iteration.