Allow generic parsing of Tasks
This PR introduces the concept of a GeneralAsanaTask, which allows the
user to provide their own parsing of the response from a Task-returning
endpoint. This is because all of the fields of a Task are not returned
by default, but can be requested. An instance of GeneralAsanaTask has
been added for Named, and a new one is on the way.

For more info, see the asana API docs on tasks:
https://developers.asana.com/docs/task