Sort planning-poker tasks by base project order
Background: When going through a planning poker session, we lose the
ordering of the tasks with respect to the base iteration (Iteration XXX)
and are forced to traverse the tasks in the order they show up in the
team iterations (RXXX, GXXX). This is a pain because it means that
sometimes the context from related tasks is missed as we are going
through the tasks, and we have to jump around.

This PR adds sorting by the base iteration field's ordering. In
order to do so, we have to provide an extra `project` parameter for the
base project. The way this is done is by first fetching all of the tasks
& projects associated with the base iteration, then filtering down to
only those within the team iteration. We then perform a secondary sort
by priority just in case the sections are out of order (this
functionality already existed, so why not keep it).