Split carryover field only in tooling
This commit changes `close-iteration` to work only on the carryover
field in asana, but splits that carryover depending on the specific
_section_ it is in. In order to accomplish this, we now operate on the
top-level project (e.g. Iteration 152) with project id and supply the
name of the subproject to filter to for stat gathering.

For example,

```
stack exec -- close-iteration --project <project id> --subproject R152
```

Will provide stats for R152, provided that <project id> matches up with
Iteration 152.

For curriculum iterations , subproject need not be supplied, so it is
left as an optional argument.