Properly filter down to only current-iteration's carryover section
If carryover has already been added to the carryover section for the
next iteration, it will currently be considered as carry-in for the
current iteration. This commit allows filtering down to the exact
carryover section within the current iteration to determine carryover to
avoid that issue. Now it doesn't matter when carryover is added to the
next iteration - if it is not carry for the current iteration, it will
not be considered carry-in.