Move library functions away from AppWith type
Rather than relying on a concrete App type, we can use a Has class to
look for the API key. This means we can re-use these library functions
in other applications with their own concrete App types.

This pattern is how almost all RIO-aware functions are written. You can
tell we're following things correctly when the only imports of our App
module are in the executables.