Use iso8601-time
This library is so much easier to use (just look at the diff) and works
the same across all GHCs we test with. The stuff in base (and so RIO) is
a mess to work with and has even been deprecated in nightly (prompting
this change). It seems the deprecation is pushing users to a more
iso8601-time like interface at least.