Migrate away from RIO, add Blammo, use extracted API client
We were using `RIO` here for two reason:

- Its Prelude
- Logging

At the time, `RIO` was a contender to replace `Freckle.App` and solve
our custom Prelude needs. This has not materialized, and we've moved
further and further away from it. The current approach is
`Freckle.App.Prelude` and a new Structured Logging library, Blammo.

This commit migrates to Blammo, but not `freckle-app`. The reason I did
not bring in freckle-app is that this project has some non-standard
behavior w.r.t. to logging configuration, which I wanted to preserve for
now. Making a project-local Prelude worked well enough.

I also extracted the API client bits to the `asana-hs` repository and
released that as the `asana` package on Hackage. I decided the api
client was more publicly-valuable, so I gave it the name "asana" and
renamed this package to "asana-cli" to avoid conflict. I plan to rename
this repository to match, but I'm not married to this name in
particular, so please speak up if you have thoughts.

(Note that naming the api client repository "asana-hs" was _not_ to
avoid a conflict, but because that's our convention for libraries with
generic names that may exist in many languages.)

With Blammo, the logs while running the CLI tool now look like this:

![](https://files.pbrisbin.com/screenshots/screenshot.2439692.png)

And if we ever try to put this in a deployed service, `LOG_FORMAT=json`
is now available.

![](https://files.pbrisbin.com/screenshots/screenshot.3235920.png)